with a bug for test

+++ main.go
@@ -1,8 +1,8 @@
 package main
 
-import "github.com/wanhanbo/GomodCase/v2/utils"
+import "github.com/wanhanbo/GomodCase/v3/utils"
 
 func main() {
-	utils.SayOne(2)
+	utils.SayOne(float64(2))
 
 }
